first part tests + norminette

+++ ft_atoi.c
@@ -6,34 +6,34 @@
 /*   By: paminna <paminna@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/02 20:15:54 by paminna           #+#    #+#             */
-/*   Updated: 2020/11/04 18:01:44 by paminna          ###   ########.fr       */
+/*   Updated: 2020/11/08 16:22:28 by paminna          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+int				ft_atoi(const char *str)
 {
-    int i;
-    long long res;
-    int sign;
+	int			i;
+	long long	res;
+	int			sign;
 
-    sign = 1;
-    i = 0;
-    res = 0;
-    while (str[i] == 32 || (str[i] > 8 && str[i] < 14))
-        i++;
-    if (str[i] == '-' || str[i] == '+')
-    {
-        if (str[i] == '-')
-            sign = -1;
-      i++;
-    }
-    while (str[i] >= '0' && str[i] <= '9')
-    {
-        res = res * 10 + str[i] - '0';
-        i++;
-    }
-    res = res * sign;
-    return (res);
+	sign = 1;
+	i = 0;
+	res = 0;
+	while (str[i] == 32 || (str[i] > 8 && str[i] < 14))
+		i++;
+	if (str[i] == '-' || str[i] == '+')
+	{
+		if (str[i] == '-')
+			sign = -1;
+		i++;
+	}
+	while (str[i] >= '0' && str[i] <= '9')
+	{
+		res = res * 10 + str[i] - '0';
+		i++;
+	}
+	res = res * sign;
+	return ((int)res);
 }
